Fix and wrapper for FFTW handle
