+++ fft/src/KokkosFFT_Host_transform.hpp
@@ -9,40 +9,40 @@
 
 namespace KokkosFFT {
 namespace Impl {
-template <typename PlanType, typename... Args>
-void exec_plan(PlanType& plan, float* idata, fftwf_complex* odata,
-               int /*direction*/, Args...) {
-  fftwf_execute_dft_r2c(plan, idata, odata);
+template <typename ScopedPlanType>
+void exec_plan(ScopedPlanType& scoped_plan, float* idata, fftwf_complex* odata,
+               int /*direction*/) {
+  fftwf_execute_dft_r2c(scoped_plan.plan(), idata, odata);
 }
 
-template <typename PlanType, typename... Args>
-void exec_plan(PlanType& plan, double* idata, fftw_complex* odata,
-               int /*direction*/, Args...) {
-  fftw_execute_dft_r2c(plan, idata, odata);
+template <typename ScopedPlanType>
+void exec_plan(ScopedPlanType& scoped_plan, double* idata, fftw_complex* odata,
+               int /*direction*/) {
+  fftw_execute_dft_r2c(scoped_plan.plan(), idata, odata);
 }
 
-template <typename PlanType, typename... Args>
-void exec_plan(PlanType& plan, fftwf_complex* idata, float* odata,
-               int /*direction*/, Args...) {
-  fftwf_execute_dft_c2r(plan, idata, odata);
+template <typename ScopedPlanType>
+void exec_plan(ScopedPlanType& scoped_plan, fftwf_complex* idata, float* odata,
+               int /*direction*/) {
+  fftwf_execute_dft_c2r(scoped_plan.plan(), idata, odata);
 }
 
-template <typename PlanType, typename... Args>
-void exec_plan(PlanType& plan, fftw_complex* idata, double* odata,
-               int /*direction*/, Args...) {
-  fftw_execute_dft_c2r(plan, idata, odata);
+template <typename ScopedPlanType>
+void exec_plan(ScopedPlanType& scoped_plan, fftw_complex* idata, double* odata,
+               int /*direction*/) {
+  fftw_execute_dft_c2r(scoped_plan.plan(), idata, odata);
 }
 
-template <typename PlanType, typename... Args>
-void exec_plan(PlanType& plan, fftwf_complex* idata, fftwf_complex* odata,
-               int /*direction*/, Args...) {
-  fftwf_execute_dft(plan, idata, odata);
+template <typename ScopedPlanType>
+void exec_plan(ScopedPlanType& scoped_plan, fftwf_complex* idata,
+               fftwf_complex* odata, int /*direction*/) {
+  fftwf_execute_dft(scoped_plan.plan(), idata, odata);
 }
 
-template <typename PlanType, typename... Args>
-void exec_plan(PlanType plan, fftw_complex* idata, fftw_complex* odata,
-               int /*direction*/, Args...) {
-  fftw_execute_dft(plan, idata, odata);
+template <typename ScopedPlanType>
+void exec_plan(ScopedPlanType scoped_plan, fftw_complex* idata,
+               fftw_complex* odata, int /*direction*/) {
+  fftw_execute_dft(scoped_plan.plan(), idata, odata);
 }
 }  // namespace Impl
 }  // namespace KokkosFFT
